Vim: Use background (headless) Make! (vim-dispatch)
Map <F5> to :Make! Instead of :Make (the previous one).

This will replace the old behavior that a splitted pane on which
the make process would run is spawned; in a new window (in tmux)
will be created in order to make the progress run asynchronously.
To see the result (stdout) in the quickfix buffer, press <F6>.
Along with these, the old behavior is mapped to <leader><F5>.

To the best of my knowledge, current there is no way to attach
custom callback when the asynchronous process ends; I have to
further investigate into this issue for better notification
(e.g. in case of failure, display the error message, etc.)