Tmux: Use updated version of 'vim-tmux-navigator' settings
For some reasons, shortcuts for vim-tmux-navigator are not working.
To adress this issue, I have updated the tmux keybinding settings
of C-{h,j,k,l,\} as specified in the updated version of the plugin.

In particular, for <C-l>, the behaviour is slightly modified:
navigate to the right tmux pane if on vim, or clear the screen otherwise
(e.g. on tmux shell) rather than navigating to the right pane.