Vim: Fix imap for arrow keys not to conflict with omnicomplete
In omnicomplete mode, when we press <up> or <down> keys which
are mapped to <c-o>{gk,gj}, the cursor will not move and
the characters 'gk' and 'gj' will be inserted literally.

This is due to a key conflict with omnicomplete; to workaround
the issue, imap key mappings for <up> and <down> are fixed.

Reference: http://vi.stackexchange.com/questions/3559/imap-beginning-with-c-o-interferes-with-omnicomplete