Zsh: Add envvar 'FZF_CTRL_R_OPTS' for fzf
The previous default configuration in fzf's <Ctrl-R> command
(for searching through command history) was without the sorted mode.
To hit the best match to the query, we had to toggle the sort mode
by pressing another <CTRL-R>, which is pretty unhandy.

A recent version of fzf started to support 'FZF_CTRL_R_OPTS',
thereby allowing to pass the '--sort' flag to fzf invoked with
<Ctrl-R> by default, so that we can easily look up the desired command.

References: junegunn/fzf#526