Zsh: Revive the VIRTUALENVWRAPPER_PYTHON variable
In commit fb6bf854, I removed the variable since I wanted to use
homebrew-shipped python for virtualenvwrapper commands (e.g. workon).
However, it turns out that in certain systems including Ubuntu Linux
python might not be located at '/usr/local/bin/python', which results
that virtualenvwrapper cannot be localized and sourced if we are
inside in a virtualenv.

A good solution to this would be to dispatch the configuration
according to the platform, i.e. use homebrew-shipped python for OS X
(Darwin) and the default system python for other platforms.