Vim: Add command 'Makeprg' to easily configure makeprg
(Example Usage) :Makeprg python main.py --with-options

The makeprg option specifies the command to run for make (<F5>),
which is, however, somewhat difficult to configure if the command
line becomes long and contains multiple words.

For 'set makeprg', we need to escape all the whitespaces with '\'
(e.g. set makeprg=python\ main.py\ --with-options); an alternative
would be using the let command (e.g. let &g:makeprg='python main.py')
which still comes to me not that handy.