Zsh: Temporary fix for the $PATH order problem
In the login shell, the order of $PATH might be different than
expected (e.g. re-sourced or interactive shells); the path entry
'~/.local/bin' did not locate at the top priority.

This is because of the order of sourcing zsh startup files
(zshenv -> zprofile -> zshrc). As /etc/zprofile in macOS contains
path_helper, the order of $PATH set in ~/.zshenv is corrupted.

To work-around this issue, the path configuration is repeated
in the ~/.zprofile, so that the ~/.local/bin/ entry comes first
also in the login shell.