Vim: Use 24-bit true color (termguicolors) whenever possible
Whenever possible, use termguicolors features to enable 24-bit
true color, even in the terminals like GUI.

There is a subtle challenge as some shell or terminal emulators
may not support 24-bit colors. We assume that modern terminal
emulators we would use will support it, or even fall-back to
256 colors nicely, but otherwise it may look broken as in mosh.

For mosh, to workaround this issue until mosh supports true color,
we detect whether the vim is running in a mosh environment.
This detection should be robust even when it is running under a
tmux session; to this end, the 'is_mosh' script is introduced.
Whenever the vim is initialized or asked to redraw the screen,
we detect mosh under which the feature shall be disabled.