zsh: Enable the node module if nvm is installed in local
In some environments where nvm is installed in local as expected
but system-wide node is not available, nvm was not being load.
Now nvm can be loaded so long as it is present.

Another effect of having this line is that nvm will no more
automatically append some lines for sourcing nvm into zshrc, which
I don't want to happen (superseded by the node module with zplug).