neovim: Auto-install neovim python package on nvim startup
Neovim's python3 should be the python in the current environment
(venv, conda, or local). If the neovim python package is missing
in the current python environment, just automatically install it
on neovim startup. It will be more helpful for most normal users.

Previously it was falling back to the alternative python (e.g.
system python) that has neovim package.

Related commit: b098e65