vim: Add plugin 'coc-fzf'
This plugin looks still incomplete but provides additional fzf-based
commands for handle coc.nvim, so having it would be a plus.

It is worth noting that this changes the behavior of :PreviewDefinition,
which previously would open the built-in :CocList but now would open
:CocFzfListLocation in a floating fzf window (which is way better).
This is because this plugin implements an event listener for
`CocLocationsChange`.