zsh: Add function 'ghad'
`ghad` stands for git history --all in a daemon mode.

It runs `git history --all` but whenever there is a change on git branch
or refspec is made, it refreshes the output (IO-notified by fswatch).

This is useful for showing up-to-date git history and graph while
performing complex git operations on shell.