vim: `<leader>cd` changes the CWD to the project root
The keymap `<leader>cd` (`:CDRoot`) will now make vim change the
current working directory to the project root for the current file or
buffer. This would be useful for following grep or fzf commands.

The previous behavior was to chdir to the base directory of the file,
which can be still called through `:CD`.