zsh: Make <CTRL-X><CTRL-E> work both in viins and vicmd mode
The keybinding <CTRL-X><CTRL-E> opens an external editor (vim)
for editing the current command line. This is working well on
vicmd mode (configured by prezto) but was not really in viins mode.

So, a custom keybinding config is added to make <CTRL-X><CTRL-E>
work well in the viins mode as well.