zsh: Use prompt plugin 'powerlevel10k'
Powerlevel10k is a fast, full-featured prompt theme and plugin for zsh.
We will be replacing the existing pure theme with a powerlevel10k
configuration that should look similar to my existing prompt theme.

The prompt theme is based on p10k's classic theme (p10k-classic),
but will be configured a lot in the following commits.