vim: Allow bang option for `:T`, `:FloatermToggleOrNew`
Example:

- `:T! ls`
- `:FloatermToggleOrNew! ls`

voldikss/vim-floaterm#98 has changed the behavior of floating terminal
to a way that the terminal buffer window won't be autoclosed.

However, I would want to keep the previous behavior (i.e., run the
command in a shell and keep the shell process running afterwards)
but only with another non-default option. That is the bang option:
for example, `:T ls` will execute the command and the process will
terminate (similar to `:terminal ls`), but `:T! ls` will run the command
as the first command typed in an interactive &shell.

One limitation is that in the latter case the floating terminal will
still not be autoclosed. This should be improved later on (requires
upstream fixes; see voldikiss/vim-floaterm#98).