vim: Make &scrolloff preserved after leaving terminal
A follow-up for 5c79586 where the scrolloff option was set to 0
because of flickering issues on terminal windows.

Until neovim 0.5 (or nightly as of Mar 17, 2020) scrolloff is a global
option, so it was resetting the scrolloff option for all normal windows.
A workaround is to use TermLeave autocmd to get the option restored.