vim: Force refresh fugitive when staging via gitgutter
Fugitive is not able to reload git index buffers upon external changes
(see tpope/vim-fugitive#1517). As a workaround, we make use of autocmd
provided by vim-gitgutter (GitGutterStage, airblade/vim-gitgutter#698)
to listen to GitGutterStage events in order to refresh fugitive buffers.

A handy utility function/command `WinDo` is also added, which is almost
same as `:windo` but preserves the current window, cursors, and view.