vim: Use expandcmd to expand makeprg for python make plugin
Similar to the way :make expands the makeprg command (e.g. `%`),
the `<F5>` plugin will send expanded commands to the terminal.
For this, a backport of `expandcmd` (named `ExpandCmd`) is imported.

When running on a test code, the command will automatically run
the class/function on which the cursor is.