vim: Set minwidth for airline coc_status/branch section
Airline's coc_status and branch are quite lengthy; when there are
a lot of text to show in the statusline, the buffer's filename might
be truncated or even hidden. Meh...

To work around this, we can responsively hide these less-important
sections by setting the minimum window width. The numbers might not
ideal but should work reasonably well, at least better than nothing.