vim: Fix a bug where floaterm window gets hidden after toggling
FloatermToggle (:T) was broken because the floating window would be
automatically hidden immediately when being toggled to show again,
as a side effect of our custom autohide functionality.

This is because an BufLeave event gets triggered while re-opening the
hidden floating window (during creation of a border window). Our fix
to this is (i) to bind the BufLeave autocommand after re-opening is
complete (when FloatermOpen is trigerred), (ii) and to clear the
autohide event handler properly.