vim: Switch to 'nvim-lsp-installer' from 'nvim-lspinstall'
`nvim-lspinstall` works okay but it lacks a feature to manage and
automatically install lsp servers. Specifically, installing multiple
LSP servers at once is very difficult and it does not provide an UI
and lua API that allow as much control as I needed.

`nvim-lsp-installer` provides command `:LspInstall` but it is much
better (in terms of UI and lua API). The new config allows us to
automatically install many LSP servers on startup so smoothly.