vim: Configure lsp-status to expose the current function name
The name of the current function under the cursor can be retrieved
by attaching the LSP client to lsp-status. This would be useful for
running unit tests, specific suites only under the cursor.

Treesitter or some dedicated plugins would be a better one to implement
such a feature, but we leave this as a future work.

Ref: #27