vim: Improve ForcePlugURI to handle divergent git trees
When changing the url of plugin repository due to fork for instance
via ForcePlugURI, vim-plug still may not update the plugin properly
if the new repository's git tree has diverged from the old one
(i.e., non-fast-forward). This happened in lualine.nvim's case.

To deal with such a case, we reset to a ancestor commit for
fast-forward to be possible after updating the plugin.