vim: Use 24-bit GUI color regardless of mosh
Previously we disabled 24-bit colors (termguicolors) when the current
session is running under mosh, because the stable release (1.3.2) of
mosh does not support true colors.

However, I will be using HEAD versions of mosh on server and client.
Therefore 24-bit colors will be enabled even if the session is running
under mosh (as well as in normal ssh sessions).

If you see any weirdness in color, you will need to reinstall mosh
with the devel (HEAD) version.