vim: Improve logics for ensuring pynvim package
When installing the pynvim package, the `--user` flag in the pip install
command would install the pynvim package (and its dependencies such
as greenlet, etc.) into $HOME/.local/lib/. This can be problematic when
different architectures are used, i.e., arm64 and x86_64 (on M1 mac),
because greenlet won't be an universal binary.

Therefore the `--user` flag should be avoided so that pynvim can be
installed on the default site-packages directory, except for when the
host python is a system-wide global one (on linux).