vim: Make sure treesitter updates parse tree when text changes (#32)
In the current implementation of treesitter, the parse tree for the
current buffer would not be automatically updated/refreshed when
text changes, unless treesitter-highlight module is enabled.
However, I don't want to use treesitter's highlight/syntax feature yet
due to some issues and conflicts.

To make the parse tree up-to-date as text changes, parser:parse()
needs to be explicitly called in the absence of treesitter-highlight.

Ref: nvim-treesitter/nvim-treesitter#2492