vim: Add plugin 'pretty-fold.nvim' to improve fold appearance
This is a great plugin that provides much better appearance of
folding. We can put the preview foldtext into the exactly same position
as when were unfolded, which provides more seamless and smooth reading.

Specifically, we can have a pretty preview of folds like

    def foo(self):
        self.bar(...) ──────────────────────── 10 lines ───

instead of (neovim default):

    def foo(self):
    +-- 10 lines: self.bar( ·······························

This plugin also provides a feature of previewing folded regions.
Use `zp` to open a floating preview window for the current fold region.