vim: Improvements to the ':F' command (FzfSmart)
The `:F <args>` command can take an optional argument.

If it is a directory, either :Files or :GFiles (if within a git repo)
on that directory will run. If it is a file, either :Files or :GFiles
will run on the directory in which the file is contained.

If no arguments are provided, it will work for the CWD directory.

Some useful commands to remember:

:F.  => Search files in the current working dir (or the git repo)
:F%  => Search files relative to the current file (or its git repo)

In addition,

:FC  => This is an alias for `:F ~/.dotfiles`.