vim: Add completion for :Pyedit commands
We use jedi completion engine to provide command-line autocompletion
for the :Pyedit, :Pysplit, and :Pyvsplit commands.

In addition, we can also jump to the python module file *without*
actually importing the module into the host python (which is slow,
and may often fail) thanks to jedi's inference engine. This will make
the speed of :pyedit commands even faster.