vim: Improve conditions for automatic insert mode on TermOpen
When a terminal buffer is opened (TermOpen), for example, by `:term`,
we would want to automatically enter the insert mode. However,
the existing autocmd may result in the non-terminal buffer unwantedly
entering the insert mode when a new terminal buffer is opened on a
floating window. To prevent this, we can enter the insert mode
only if the current window has the terminal buffer being opened.

See rcarriga/neotest#2 for reference.