vim: Use plugin 'nvim-ufo' for advanced folding
nvim-ufo is a great plugin that allows customization of folding,
including foldtext with "highlight" support (via virtual text).

I use a custom virtual text handler to customize the foldtext
to my taste. This includes (i) the last line of the folded region
(but up to a single token), and (ii) summarization of the folded region
around the '⋯' ellipsis.

nvim-ufo provides tressitter/lsp fold provider. To me they appear
to be a bit unstable (e.g., some bugs where folds get unwantedly
open or closed upon saving, fold do not update correctly).
In such cases we could use a 'null' provider as a workaround.