zsh: Migrate from zgen to antidote
zgen is no longer maintained; but it has a little bit of bug.
We switch to another modern, well-maintained zsh plugin framework
called `antidote` to list and manage zsh plugins.

For the time being we use 'dynamic loading' (it's only a bit slow), but
later 'static loading' should be used for a better startup performance.