zsh: Automatically determine #lines for git history aliases (gh)
Instead of having a fixed number of lines (or log entries), say -n30,
to be printed in `git history` (`gh`, etc. aliases), we can determine
the number of lines and the height to be truncate at w.r.t the current
terminal size. This will make the output of `gh`, `gha`, etc. always
fit all in the current terminal whatever size it has, not entering
the "less" `lines .../... (END)`.