vim: Show details for completion entries such as type info (#39)
Some language servers provide a "detail" for a completion entry, e.g.,
type information. If this is available, we can show it in "item.menu"
instead of the name of LSP server source.

Python LSP servers (pyright), however, do not have this field; for
some specific cases like auto-import, we can use some LSP-specific
fields to resolve the detail. For UltiSnips snippets, this can be
the description of the snippet, if any.

Detail information for the completion items will be highlighted in a
different color (CmpItemMenuDetail).