vim(lsp): Reload and attach LSP on startup and LSP installation
One problem of mason.nvim config:

mason.nvim would install packages (LSP servers) asynchronously,
so lspconfig[...].setup {...} calls won't be made for the LSP servers
that are not installed by vim startup (which will be automatically
installed after entering vim). This is a bit inconvenient because one
would need to restart vim in order for the new LSP servers to start.
Also, even if LSP servers are installed, buffers that are open prior
to execution of config/lsp.lua won't have LSP clients attached.

My solution to these issues is:

- Reload and attach LSP after sourcing the config is done, by firing
  a BufReadPost event for each of the (already) open buffers.

- When a new LSP server is installed through mason, we execute a
  callback so that lspconfig can be set up *after LSP installation*,
  followed by reloading of LSP.

This way, I can have the same behavior as nvim-lsp-installer that
we previously had..