vim(plugins): Make ForcePlugURI work lazily
The command ForcePlugURI is used to fix URI of the plugin repository
whose URI has changed (or used when we switch to a different fork),
but checking git remote url was done every time vim is starting.
This is inefficient as it may waste some time on vim startup; instead,
we defer checking of git remote URIs to when PlugInstall or PlugUpdate
is actually called.