vim(plugins): Remove the use of HasPlug and IsPlugInstalled
HasPlug and IsPlugInstalled are helper functions to conditionally
activate plugin configs defined in vimrc. As a step towards migrating
to the lazy.nvim plugin manager, we refactor the plugin config chunk
as a separate (vimscript) function, which will be executed at the
end of vimrc execution when vim-plug is used.