vim(plugins): Lazy-load plugins as much as possible
To improve neovim startup time, we lazy-load as many plugins as possible.

Some plugins can be loaded on demand upon commands or keymaps, and
others can be deferred to after UIEnter. Lua-based neovim plugins
that are configured in config/*.lua can be simply set as lazy plugins.

There are some tricky cases where plugins are not working well when
lazy-loaded, which are left with eager loading. Or some kind of hacks
will be needed (e.g., fugitive, wilder, semshi, etc.)

With this, vim startuptime has been improved:

- `:Lazy profile` IMPROVED: ~50ms,  BEFORE: ~120ms (more accurate measure)
- `:StartupTime`  IMPROVED: ~100ms, BEFORE: ~160ms