vim(plugins): Refactor lua config modules to be invoked by lazy.nvim
Another follow-up step for #44 is to remove all the manual sourcing
of `config/*` lua scripts from init.nvim and have these plugin-specific
configs sourced on demand upon (lazy) plugin loading.

Config scripts are refactored to be modular in the setup{ ... }-style,
and we also make them be re-sourceable as well.