vim(lsp): Fix performance regression on startup + lazy lsp loading
Rarely, a FZF window or some other windows in the UI was freezing
and blocked LSP-related plugins are lazy loading (upon LspSetup).

This is probably due to the expensive BufReadPost events taking place
in reload_lsp(); we can have only the lspconfig autostart autocmds
executed when reloading LSP for the existing buffers.