vim(options): Set mouse=a (all)
We enable vim's mouse support in the insert mode as well. The previous
option set mouse=nvc (excludes insert mode) was there for many years
for a historcal reason -- to disable mouse while grabbing some text.
In modern terminal emulators (iTerm, kitty, etc.) this is supported
by holding some modifier keys (shift or alt) to temporarily disable
vim's mouse support. So as of now, there should be no reason to not use
the mouse support in the insert mode as well.