vim(treesitter): Automatically fix 'invalid node type' parser error
Whenever nvim-treesitter updates to a different version, all the
compiled treesitter parsers must be invalidated and re-built to prevent
the annoying 'query: invalid node type' errors on treesitter parsers
(see nvim-treesitter/nvim-treesitter#3092). Despite post-update hook
of the plugin, this broken state may happen due to possible manual
configuration of the plugin and neovim versions.

An user-friendly, fool-proof fix to this would be to try detecting and
automatically fixing this (quite a common) situation. We call a custom
function _G.TreesitterParse on text edit or after reading the buffer,
in which we can catch the treesitter parser error and force-reinstall
treesitter parsers.