vim(lsp): Force enable neodev plugin support
If the lsp file that triggered an initialization of lua_ls LSP server
is outside the vim config or is a single-file, neovim API won't be
added to the library path of the lua_ls LSP server.

This is a bug of neodev (see folke/neodev.nvim#158); as a workaround,
we can enable neodev's plugin support, which manipulates lua_ls's
library path, when the root_dir is located under dotfiles.