vim(lsp): Improve choice heuristic among multiple defs on PeekDefinition
When there are multiple LSP definitions found for PeekDefinition, we
used to choose the first result entry. However this may not give the
best result for some parameters or fields, because it may give a trivial
result peeking the same file.

One heursitic we can have here is to prefer definitions defined in "far"
location, i.e., in a different file than the requested (current) buffer.
If all definition results are coming from the same file, of course the
(usually the only one) definition should be peeked.