vim(options): Increase &updatetime to 1000 (ms)
This is the timeout value used for CursorHold event. After &updatetime
milliseconds, LSP diagnostics popup will show (if any).

The value was previously set to 200ms so as to make gitgutter more
responsive (see e7bb502b), but for LSP diagnostics this feels too short
as diagnostics popup window may show up quite aggressively. For LSP,
1000ms would be a better value. We no longer use gitgutter but instead
we use gitsigns, which does not depend on &updatetime and is able to
update signs almost immediately, so now there's no point of having
short &updatetime.