vim(lsp): Fix performance regression on UI freezing during lsp init
In neovim nightly (v0.10.x), the UI would freeze with high CPU usage
when some language servers (e.g., python; pyright) are launching and
attaching to a buffer for the first time.

This is due to the new default LSP capabilities setting introduced in
neovim/neovim#23190. It was fine in neovim v0.9.1 stable.

The `workspace.didChangeWatchedFiles.dynamicRegistration = true` option
allows the LSP server to be notified with file changes on disk, but
this results in scanning the big filesystem (somehow in a blocking or
synchronous manner?). Until further follow-up and performance
improvement is made (see neovim/neovim#23291), we will disable this
client capabilities options to make neovim not freeze.