vim(tabline): Make tabline not flicker when on vim startup
We can simplify settings for vim-xtabline now we require neovim >= 0.8;
there is no need to change the XTabMode after VimEnter (which causes
some kind of UI flickering due to the delayed execution).

This also removes the 'TablineHybridMode' config which is no longer
used, where the tabline mode can change automatically depending on
the number of tabs opened.