vim(lsp): Try loading telescope on on_attach() for jump keymaps
If telescope is not lazy-loaded yet by the time on_attach() is called,
keymaps like `gd`, etc. will resolve to the original lsp commands.
At the cost of loading telescope (it's only 5 ms) just slightly early,
we can always ensure that those jump keymaps will work with telescope
as long as the plugin is installed.