vim(treesitter): Remove treesitter parsing trigger workaround
Remove the workaround introduced in ec2f2529 (#32) when neovim was ~0.7.
This was needed because treesitter's incremental parsing did not happen
when the highlight module was not enabled. This might have been fixed
by neovim core (for 0.8+) or by the nvim-treesitter plugin (not sure),
but incremental parsing now works OK even when treesitter highlight
is not enabled, so we probably no longer need this workaround.