vim(python): Remove python-mode and py-enhanced-syntax
python-mode has been progressively deprecated in favor of LSP, linter,
and many other modern tools in neovim. It's finally the time to drop
the python-mode plugin (which is very heavy!).

We have been disabling debugging, folding, rope(refactoring), renaming,
linting, and preview, so these can be safely removed.

indent: The PEP-8 compatible indentexpr would be the only useful stuff
  for me among its feature set. Indent functions will be cherry-picked.

syntax/highlighting: We are moving python highlighting and syntax
  support to treesitter at least for python, so we can also remove
  the vim-python-enhanced-syntax plugin.

keymap: One more minor thing is a keymap to toggle the breakpoint
  support, which can be implemented separately in the subsequent commits.