vim(treesitter): Auto-recover parser errors if treesitter.start fails
It is reported that when updating dotfiles and nvim-treesitter plugins
since a long time ago where parsers were quite outdated, an annoying
error of "invalid node type" can happen and nvim can even freeze.

A solution to this is to run the automatic treesitter recovery process
(:TSUpdateSync) when vim.treesitter.start() has failed.