zsh: Fix a possible bug where zsh prompt redirected to ANTIDOTE_BUNDLE
It's not easy to 100% reproduce this, but in some rare cases zsh prompts
may be redirected and written to ANTIDOTE_BUNDLE (~/.zsh_plugins.zsh).
Possibly a bug of zsh, or a bug of antidote, or both; but using
a subshell when writing into the bundle file simply fixes the bug.