zsh: Do not add {miniforge,miniconda}/bin/ to $PATH, but add condabin/
anaconda (miniforge/miniconda) `bin/` directories should not be added
to `$PATH` on a startup shell; this results in choosing a wrong python
interpreter unless the conda environment is explicitly sourced.

To purpose of scanning conda directories is to enable the `conda`
command (from any base conda environment), so only `condabin/` should be
added instead.

Fix-up to #49.