vim(lsp): pyright should always use the current python on $PATH
It seems that in some cases pyright picks up a different python than
the current python in accordance with $PATH (i.e., `which python3`),
e.g., when activated conda or python virtualenvs. It might be a bug
because pyright's [import-resolution] doc says (in 3.) that the default
python would be used when determining the python environments.

We enforce pyright to always use the same python as the rplugin host
and the python on $PATH by setting the `python.pythonPath` config.

[import-resolution]: https://github.com/microsoft/pyright/blob/main/docs/import-resolution.md#configuring-your-python-environment