vim(build): Use `<F5>` for `:Build` and `<M-F5>` for `:Debug`
From now on, `<F5>` will be mapped to `:Build` rather than `:Debug`
which will result in less confusions. Although it's a bit different from
the VS code convention, building/running programs in a non-debugging
mode would be more common and hence the go-to default for my usage.

Programs can run in a debugging mode with `<M-F5>` or `:Debug`
(when the `dap` branch gets merged in the near future).