vim: Fix erroneous startinsert on new windows jumped from terminal
There is a bug (or just a strange behavior?) of neovim where buffer
options (e.g. &buftype) gets updated late on WinEnter autocmds.

When a new window is opened, the new buffer that will be opened in the
new window will take effect *after* the WinEnter autocmd. The actual
buffer would not yet have been loaded at the time when autocmd WinEnter
events are being executed. As a result, the new window will enter the
insert mode even if it's not having a terminal buffer.

One known solution is to check if <amatch> is empty.

Credit: https://vi.stackexchange.com/questions/15966/strange-behaviour-of-autocmd

See also: 0fef0f059 and e684e2258 (the same bug, just workarounds)