tmux: Adjust status-right-length using a hook to deal with "not ready"
Problem: When attaching to a tmux session, tmux prints an excessive
message saying `statusbar.tmux` is not ready, which overflows in the
tmux statusline. This is because the CPU/RAM component takes around
1.0 seconds until it can print the output.

Solution: Hide status-right by setting status-right-length = 1 using a
client-attached hook. The status-right-length can be adjusted back to
the normal limit after around 1 seconds.