vim(formatting): Configure formatexpr to use conform formatters
By default neovim will configure (buffer-local) formatexpr with LSP
backend. Similar to LSP, we register the buffer-local formatexpr for all
filetypes that are configured with conform formatters.

This allows `gq` commands as well as automatic line wrapping in the
insert mode. As of v4.2.0 conform fall backs to the built-in formatter
(i.e. line wrap) when neither conform or LSP formatters were available,
but this leads to messing up existing formatting (especially when it
comes to range formattings); see stevearc/conform.nvim#55. As a
workaround, we should NOT fallback to the built-in formatexpr if the
formatexpr function was invoked outside the insert mode.