vim(git): Force-reload fugitive index buffer upon BufEnter
Fugitive git index buffers won't automatically reload upon external
changes (see tpope/vim-fugitive#1517). Gitsigns can trigger reloading
but this only works when the working copy buffer is in the "diff" mode;
which we disabled for `:GitThreeWayDiff`.

A good workaround is to reload fugitive git index when entering the
fugitive buffer.