vim(statuscolumn): Prioritize non-staged gitsigns over staged gitsigns
Since only one git sign can appear, if there are both staged changes and
unstage changes on a (same) line, showing "this line is staged" can be
confusing. This is because gitsigns currently have the same priority 6
for both types of signs. We choose to show gitsigns for unstaged changes
only, because this is often more informative.

See #51.