zsh: Fall back to TERM=xterm-256color if terminfo is not available
When terminfo is not available for the $TERM given in the environment
(e.g. TERM set in by the terminal emulator, tmux, or through SSH, but
the terminfo isn't installed on a remote server), the shell will be
broken and completely unusable. Falling back to xterm-256color would
be a good workaround.