vim(lsp): Do not enable pylint unless a project explicitly configures
pylint is quite annoying because it is slow and it often lints too
aggressively. These days, ruff serves as a good, fast alternative to
pylint even including some of pylint's rules that are useful.

Unless a python project is explicitly configured to use
pylint as a required linter (e.g., by having the `.pylintrc` file),
we now don't enable pylint by default.