vim(statusline): Move AutoBuild status icon to "section b"
Some components in the lualine "section c" might be easily hidden if the
text content overflows. An important component (:AutoBuild status) that
should be seen at all times can be moved to "section b".

Also change the emoji for the 'waiting (listening)' status due to the
unicode width issue in some terminals.