vim(neotree): Do not use custom implementation for trailing_slash
A custom implementation to draw trailing slashes for directories works
quite buggy in recent versions of neotree. The problem is: other
components to be drawn on the right side (such as git status) misalign.
For some reason, the built-in option now works well, so we switch to the
built-in implementation.