vim(lsp): Allow `python` as package path within the workspace
In addition to the default package search path:

- the root directory of the workspace, and
- the `src/` directory on the workspace,

we also allow the `python/` directory for where a package can be found
(e.g. for package `mypackage`, `$PROJECT_ROOT/python/mypackage` instead
of `$PROJECT_ROOT/mypackage` or `$PROJECT_ROOT/src/mypackage`). This is
a bit non-standard but allowable with some reasonable convention.

Note: the setuptools equivalent is:

```toml
[tool.setuptools.packages.find]
    where = ["python"]
```