vim(python): Use treesitter indent, no more vimscript PEP8 indent
The existing PEP8PythonIndent is buggy if treesitter syntax highlight is
used. We had a workaround to detect whether it's comment, but due to its
poor logic there is a huge performance drawdown when the file is too big
and the filesystem is not fast enough.

Although treesitter's python indentation also has a few bugs and
incorrectness, and it lacks a proper support of 4-space hanging indent,
using treesitter's indentation would provide overall the best user
experience (a good project would be using some good formatters like
ruff/yapf/black).