vim(snippets/python): Make debugger snippets more formatter-friendly
*Problem*:
The comment markers used by yapf and ruff-format are different and not
compatible with each other. Therefore, having a long line like

    import pudb; pudb.set_trace()  # XXX DEBUG  # yapf: disable  # fmt: off

doesn't seem good. Since yapf is deprecated in my config, we could have
a single `fmt: off` marker to prevent formatters from splitting the
snippet line into multiple lines, but there's another neat trick.

*Solution*:
Use the `__import__` builtin to make the entire debugger entry call a
single-line, single-statement code. `importlib.import_module` would be a
better API, but it requires an additional import statement.

ptpython is one exception, as there's no easy way to write it nicely.
A workaround is to use `exec()`; it's fine because it's just for
debugging.