tmux: ensure set-clipboard is on so that OSC52 can work properly
When the `set-clipboard` option is `on`, applications inside tmux (such
as neovim) are allowed to set tmux's clipboard which is then transmitted
through OSC52. This option seems to be the default, or at least should
be `external` (where OSC52 still works), but it turns out that on some
distros the option is turned off by default.

Without this option being set, OSC52 will not work properly in neovim
(try `"+y` or `,y`); often it hangs with the following message:
    Waiting for OSC 52 response from the terminal. Press Ctrl-C to interrupt …

Reference: https://github.com/tmux/tmux/wiki/Clipboard