vim(lsp): drop pyright support and fix basedpyright settings
In recent versions of basedpyright, `python.pythonPath` seems to be the
only working setting to choose which python environment to use; the
legacy setting `basedpyright.pythonPath` used to work but it doesn't
work well anymore possibly because of a behavior change.

To make basedpyright settings easier, we now drop support for the
vanilla MS pyright in favor of basedpyright. Basedpyright LSP has
been working great and superior to pyright in many ways, so it's no
longer experimental. Pyright will always be disabled even if it's
already (and previously) installed via Mason.