vim(markdown): use dedicated highlight group for inline raw code
Previously the same highlight group for raw markdown codeblock
(`@markup.raw.block.markdown`) was used for inline code, but this
highlight group defines a background color but NOT a foreground color.
This leads to a problem where inline codes in a markdown heading will
not easily visible if the heading foreground will resolve to a black-ish
color as markdown codeblocks have a very dark background.

e.g.
>  ## Heading: `invisible` visible

A solution is to use a highlight group that defines both of the fg and
bg color. There is a dedicated hlgroup `@markup.raw.markdown_inline`
defined by treesitter; it was being linked to `Constant` by default, but
now we can define a similar foreground (like `Constant`) and a dark
background color.