vim(init): reduce time on setting the colorscheme
The command 'colorscheme xoria256-wook' takes around 8~10ms (M1 mac).
This is because the command would trigger lazy.nvim to invoke a
ColorschemePre event to list all the available colorschemes (and load
if lazy loading is involved) by accessing the filesystem.

But for the startup this is not necessary, so we can have a small
performance improvement by around 10~20 milliseconds.