vim(pynvim): rewrite check/installation process to be asynchronous
Problem: On startup, pynvim config would run the external python3 to
check python3 and pynvim version. But the system() calls are blocking,
so if external python3 is very slow (e.g. the disk is slow or on NFS)
UI might noticeably freeze for a while.

Solution: Refactor the entire logic to make such checks work in an
asynchronous fashion. This would require vim.system() for the sake of
simplicity of implementation and hence neovim 0.10+.

Also simplify the implementation for installing pynvim with termopen
rather than managing the progress buffer manually.