vim(markdown): attach to unlisted buffers as well when lazy-loaded
Problem: render-markdown is lazy-loaded, so the buffers created before
loading of the plugin would not have markdown rendering attached. We had
a fix for this but this was not being applied to 'unlisted' buffers
(usually floating windows such as hovering doc, but some markdown-ish
scratch buffers such as codecompanion are also applied).

Solution: Make `rc_utils.bufdo` be able to enumerate unlisted buffers
as well (optionally). We also get the list of filetypes to attach from
the plugin configuration, so it can also work for other &filetype's
other than markdown.